# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.1"
    },
    "paths": {
        "/api/city": {
            "get": {
                "tags": [
                    "City"
                ],
                "summary": "Get a list of city with dynamic filtering, sorting, pagination, and advanced search",
                "operationId": "c2504917c4da1139bffc3dd1d5409f8f",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Dynamic filtering with multiple fields (e.g., `status:active|age:gt:30`)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Full-text search (e.g., `keywords`)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort the results by a specific field (e.g., `price` or `name`)",
                        "schema": {
                            "type": "string",
                            "example": "created_at"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Direction of sorting (asc or desc)",
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination (default: 1)",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 10)",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CityResource"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "City"
                ],
                "summary": "Store a new city",
                "operationId": "a8227d5bf6c485aeef4f403040f32ecc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "City created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CityResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/city/{id}": {
            "get": {
                "tags": [
                    "City"
                ],
                "summary": "Get a specific city",
                "operationId": "5a7b6766b61b35c52ccb43edd7a33259",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CityResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "City"
                ],
                "summary": "Update a specific city",
                "operationId": "d0af268d4633a5d5a5714dd23ecdec69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "City updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CityResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "City"
                ],
                "summary": "Delete a city",
                "operationId": "20163308af9adc11a49d350825877375",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "City deleted successfully"
                    }
                }
            }
        },
        "/api/city/{id}/restore": {
            "put": {
                "tags": [
                    "City"
                ],
                "summary": "Restore a deleted city",
                "operationId": "e3f4d1b6e94745838eeca0bb152fd57e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City restored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The record has been restored successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/city/{id}/force-destroy": {
            "delete": {
                "tags": [
                    "City"
                ],
                "summary": "Permanently delete a city",
                "operationId": "06f83877ea7e199a9b35aeb8afad2ce3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "City permanently deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CityRequest": {
                "required": [
                    ""
                ],
                "properties": {
                    "region_id": {
                        "type": "integer",
                        "example": "5"
                    },
                    "name_en": {
                        "type": "string",
                        "example": "Name_en Example"
                    },
                    "name_mm": {
                        "type": "string",
                        "example": "Name_mm Example"
                    },
                    "code": {
                        "type": "string",
                        "example": "Code Example"
                    }
                },
                "type": "object"
            },
...
            "CityResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "88"
                    },
                    "region_id": {
                        "type": "integer",
                        "example": "39"
                    },
                    "name_en": {
                        "type": "string",
                        "example": "Name_en Example"
                    },
                    "name_mm": {
                        "type": "string",
                        "example": "Name_mm Example"
                    },
                    "code": {
                        "type": "string",
                        "example": "Code Example"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-31 09:14:28"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-31 09:14:28"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token Based Authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}